refactor: move all SpontanResult database requests to SpontanResultService
