+++ src/main/kotlin/odyseja/odysejapka/dashboard/TeamsFormController.kt
@@ -3,7 +3,7 @@ package odyseja.odysejapka.dashboard
 import odyseja.odysejapka.form.TeamResultEntity
 import odyseja.odysejapka.form.TeamResultService
 import odyseja.odysejapka.spontan.SpontanResultEntity
-import odyseja.odysejapka.spontan.SpontanResultRepository
+import odyseja.odysejapka.spontan.SpontanResultService
 import odyseja.odysejapka.timetable.PerformanceGroup
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.GetMapping
@@ -18,7 +18,7 @@ class TeamsFormController(
     private val stageAccessService: StageAccessService,
     private val userAccessService: UserAccessService,
     private val teamResultService: TeamResultService,
-    private val spontanResultRepository: SpontanResultRepository
+    private val spontanResultService: SpontanResultService
 ) {
 
     @GetMapping
@@ -68,6 +68,6 @@ class TeamsFormController(
 
     private fun getSpontanResults(performanceIds: List<Int>): Map<Int, SpontanResultEntity> {
         if (performanceIds.isEmpty()) return emptyMap()
-        return spontanResultRepository.findAllByPerformanceIdIn(performanceIds).associateBy { it.performanceId }
+        return spontanResultService.getSpontanResults(performanceIds).associateBy { it.performanceId }
     }
 }
